Bug 859021 - Fix remaining MSVC build warnings and remove the MSVC exemption for FAIL_ON_WARNINGS in content/svg/content/src. r=dholbert
UltraBlame original commit: 7fe1df8b1a9b3009edc8158cb16349d5dc8b8048

+++ content/svg/content/src/SVGContentUtils.cpp
@@ -260,13 +260,13 @@ SVGContentUtils::AngleBisect(float a1, float a2)
 {
   float delta = fmod(a2 - a1, static_cast<float>(2*M_PI));
   if (delta < 0) {
-    delta += 2*M_PI;
+    delta += static_cast<float>(2*M_PI);
   }
   
   float r = a1 + delta/2;
   if (delta >= M_PI) {
     
-    r += M_PI;
+    r += static_cast<float>(M_PI);
   }
   return r;
 }
